Develop

+++ org.envirocar.app/src/org/envirocar/app/views/tracklist/AbstractTrackListCardFragment.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2013 - 2019 the enviroCar community
+ * Copyright (C) 2013 - 2021 the enviroCar community
  *
  * This file is part of the enviroCar app.
  *
@@ -38,6 +38,7 @@ import androidx.core.content.FileProvider;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
+import com.google.android.material.dialog.MaterialAlertDialogBuilder;
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
 import com.google.android.material.snackbar.Snackbar;
 
@@ -49,12 +50,12 @@ import org.envirocar.app.handler.TrackUploadHandler;
 import org.envirocar.app.handler.preferences.UserPreferenceHandler;
 import org.envirocar.app.handler.agreement.AgreementManager;
 import org.envirocar.app.injection.BaseInjectorFragment;
-import org.envirocar.app.views.utils.DialogUtils;
 import org.envirocar.app.views.utils.ECAnimationUtils;
 import org.envirocar.core.entity.Track;
 import org.envirocar.core.exception.NotConnectedException;
 import org.envirocar.core.exception.UnauthorizedException;
 import org.envirocar.core.logging.Logger;
+import org.envirocar.core.util.FileWithMetadata;
 import org.envirocar.remote.serde.TrackSerde;
 import org.envirocar.core.EnviroCarDB;
 
@@ -165,76 +166,33 @@ public abstract class AbstractTrackListCardFragment<E extends RecyclerView.Adapt
      */
     protected abstract void loadDataset();
 
-    protected void exportTrack(Track track) {
+    protected abstract void showNoTracksInfo();
+
+    protected void shareTrack(Track track) {
 
         try {
-            if (checkStoragePermissions()) {
-                // Create an sharing intent.
-                Intent sharingIntent = new Intent(Intent.ACTION_SEND);
-                sharingIntent.setType("application/json");
-                //  Uri shareBody = Uri.fromFile(TrackSerde.exportTrack(track).getFile());
-                Uri shareBody = FileProvider.getUriForFile(
-                        getActivity(),
-                        getActivity().getApplicationContext()
-                                .getPackageName() + ".provider", TrackSerde.exportTrack(track).getFile());
-                sharingIntent.putExtra(android.content.Intent.EXTRA_SUBJECT,
-                        "EnviroCar Track " + track.getName());
-                sharingIntent.putExtra(android.content.Intent.EXTRA_STREAM, shareBody);
-                sharingIntent.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
-                // Wrap the intent with a chooser.
-                startActivity(Intent.createChooser(sharingIntent, "Share via"));
-            } else {
-                requestStoragePermissions();
-            }
+            // Create export file
+            FileWithMetadata trackFile = TrackSerde.createTrackFile(track,
+                    String.valueOf(getContext().getExternalCacheDir()));
+
+            // Create an sharing intent.
+            Intent sharingIntent = new Intent(Intent.ACTION_SEND);
+            sharingIntent.setType("application/json");
+            //  Uri shareBody = Uri.fromFile(TrackSerde.exportTrack(track).getFile());
+            Uri shareBody = FileProvider.getUriForFile(getContext(),
+                    "org.envirocar.app.provider", trackFile.getFile());
+            sharingIntent.putExtra(android.content.Intent.EXTRA_SUBJECT,
+                    "EnviroCar Track " + track.getName());
+            sharingIntent.putExtra(android.content.Intent.EXTRA_STREAM, shareBody);
+            sharingIntent.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
+            // Wrap the intent with a chooser.
+            startActivity(Intent.createChooser(sharingIntent, "Share via"));
 
         } catch (IOException e) {
             LOG.warn(e.getMessage(), e);
             Snackbar.make(getView(),
                     R.string.general_error_please_report,
                     Snackbar.LENGTH_LONG).show();
-        }
-    }
-
-    /**
-     * Return the current state of the permissions needed.
-     */
-    private boolean checkStoragePermissions() {
-        int permissionState = ActivityCompat.checkSelfPermission(getActivity(),
-                Manifest.permission.WRITE_EXTERNAL_STORAGE);
-        return permissionState == PackageManager.PERMISSION_GRANTED;
-    }
-
-    private void requestStoragePermissions() {
-        boolean shouldProvideRationale =
-                ActivityCompat.shouldShowRequestPermissionRationale(getActivity(),
-                        Manifest.permission.WRITE_EXTERNAL_STORAGE);
-
-        // Provide an additional rationale to the getUserStatistic. This would happen if the getUserStatistic denied the
-        // request previously, but didn't check the "Don't ask again" checkbox.
-        if (shouldProvideRationale) {
-            LOG.debug("Requesting Storage permission. Displaying permission rationale to provide additional context.");
-
-            DialogUtils.createDefaultDialogBuilder(getContext(),
-                    R.string.request_storage_permission_title,
-                    R.drawable.others_settings,
-                    R.string.permission_rationale_file)
-                    .positiveText(R.string.ok)
-                    .onPositive((dialog, which) -> {
-                        // Request permission
-                        ActivityCompat.requestPermissions(getActivity(),
-                                new String[]{Manifest.permission.WRITE_EXTERNAL_STORAGE},
-                                REQUEST_STORAGE_PERMISSION_REQUEST_CODE);
-                    })
-                    .show();
-
-        } else {
-            LOG.info("Requesting permission");
-            // Request permission. It's possible this can be auto answered if device policy
-            // sets the permission in a given state or the getUserStatistic denied the permission
-            // previously and checked "Never ask again".
-            ActivityCompat.requestPermissions(getActivity(),
-                    new String[]{Manifest.permission.WRITE_EXTERNAL_STORAGE},
-                    REQUEST_STORAGE_PERMISSION_REQUEST_CODE);
         }
     }
 
@@ -303,20 +261,20 @@ public abstract class AbstractTrackListCardFragment<E extends RecyclerView.Adapt
                 R.id.fragment_tracklist_delete_track_dialog_trackname)).setText(track.getName());
 
         // Create a dialog that deletes on click on the positive button the track.
-        DialogUtils.createDefaultDialogBuilder(getActivity(),
-                R.string.trackviews_delete_track_dialog_headline,
-                R.drawable.ic_delete_white_24dp,
-                contentView)
-                .positiveText(R.string.ok)
-                .negativeText(R.string.cancel)
-                .onPositive((materialDialog, dialogAction) ->
-                        mBackgroundWorker.schedule(() -> {
-                            // On a positive button click, then delete the track.
-                            if (upToDateRef.isLocalTrack())
-                                deleteLocalTrack(track);
-                            else
-                                deleteRemoteTrack(track);
-                        }))
+        new MaterialAlertDialogBuilder(getActivity(), R.style.MaterialDialog)
+                .setView(contentView)
+                .setTitle(R.string.trackviews_delete_track_dialog_headline)
+                .setIcon(R.drawable.ic_delete_white_24dp)
+                .setPositiveButton(R.string.ok,
+                        (materialDialog, dialogAction) ->
+                                mBackgroundWorker.schedule(() -> {
+                                    // On a positive button click, then delete the track.
+                                    if (upToDateRef.isLocalTrack())
+                                        deleteLocalTrack(track);
+                                    else
+                                        deleteRemoteTrack(track);
+                                }))
+                .setNegativeButton(R.string.cancel,null)
                 .show();
     }
 
@@ -418,6 +376,10 @@ public abstract class AbstractTrackListCardFragment<E extends RecyclerView.Adapt
                     showSnackbar(String.format(getString(R.string
                             .track_list_delete_track_success_template), track.getName()));
                     hideProgressView();
+
+                    if (mTrackList.isEmpty()) {
+                        showNoTracksInfo();
+                    }
                 } else {
                     showSnackbar(String.format(
                             getString(R.string.track_list_delete_track_error_template),
